renamed model/Feed to model/Article
