+++ src/core/domain/model/Article/Article.ts
