+++ src/ui/Feed/__tests__/Feed.ispec.tsx
@@ -2,8 +2,8 @@ import { fireEvent, render } from '@testing-library/react-native'
 import { asFunction } from 'awilix'
 
 import { container } from '../../../core/_di'
-import { Article } from '../../../core/domain/model/Feed/Feed'
-import { anArticle, aSetOfArticles } from '../../../core/domain/model/Feed/Feed.model'
+import { Article } from '../../../core/domain/model/Article/Article'
+import { anArticle, aSetOfArticles } from '../../../core/domain/model/Article/Article.model'
 import { getArbitraryDate } from '../../../core/domain/utils'
 import { NavigationProp, RouteProp } from '../../_navigation/NavigationTypes'
 import { Feed } from '../Feed'
